Restructure auth token code and update author field usage
